搜索和删除用户
只有管理员才能搜索和删除

+++ src/main/java/com/tokey/usercenter/Controller/UserController.java
@@ -10,7 +10,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.tokey.usercenter.constant.UserConstant.ADMIN_TOLE;
+import static com.tokey.usercenter.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * 用户接口
@@ -32,8 +37,8 @@ public class UserController {
         String userPassword = userRegisterRequest.getUserPassword();
         String checkPassword = userRegisterRequest.getCheckPassword();
 
-        long id = userService.userRegister(userAccount, userPassword, checkPassword);
-        return id;
+        return userService.userRegister(userAccount, userPassword, checkPassword);
+
     }
 
     @PostMapping("/login")
@@ -50,20 +55,41 @@ public class UserController {
     }
 
     @GetMapping("/search")
-    public List<User> searchUsers(String username) {
+    public List<User> searchUsers(String username, HttpServletRequest request) {
+        if (!isAdmin(request)) {
+            return new ArrayList<>();
+        }
+
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(username)) {
             //like第二个参数默认是%  %
             queryWrapper.like("username", username);
         }
-        return userService.list(queryWrapper);
+        List<User> userList = userService.list(queryWrapper);
+        return userList.stream().map(user -> userService.getSafetyUser(user)).collect(Collectors.toList());
     }
 
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody long id) {
+    public boolean deleteUser(@RequestBody long id, HttpServletRequest request) {
+        if (!isAdmin(request)) {
+            return false;
+        }
         if (id <= 0) {
             return false;
         }
         return userService.removeById(id);
+    }
+
+    /**
+     * 是否为管理员
+     *
+     * @param request
+     * @return
+     */
+    private boolean isAdmin(HttpServletRequest request) {
+        User user = (User) request.getSession().getAttribute(USER_LOGIN_STATE);
+
+        return user != null && user.getUserRole() == ADMIN_TOLE;
+
     }
 }
